Fix redundant type converstion (#13076)
* Fix redundant type converstion

* Revert generated changes

---------

+++ beacon-chain/state/state-native/setters_attestation.go
@@ -54,7 +54,7 @@ func (b *BeaconState) AppendCurrentEpochAttestations(val *ethpb.PendingAttestati
 	}
 
 	atts := b.currentEpochAttestations
-	max := uint64(params.BeaconConfig().CurrentEpochAttestationsLength())
+	max := params.BeaconConfig().CurrentEpochAttestationsLength()
 	if uint64(len(atts)) >= max {
 		return fmt.Errorf("current pending attestation exceeds max length %d", max)
 	}
@@ -84,7 +84,7 @@ func (b *BeaconState) AppendPreviousEpochAttestations(val *ethpb.PendingAttestat
 	}
 
 	atts := b.previousEpochAttestations
-	max := uint64(params.BeaconConfig().PreviousEpochAttestationsLength())
+	max := params.BeaconConfig().PreviousEpochAttestationsLength()
 	if uint64(len(atts)) >= max {
 		return fmt.Errorf("previous pending attestation exceeds max length %d", max)
 	}
